All Netchdfs have chunkIteration. sections < varshape not yet supported. Striders need to be longs. HDF5 chunkIteration gets 156 Mb/sec. 1.86x calls to Clib.

+++ core/src/main/kotlin/com/sunya/cdm/layout/MaxChunker.kt
@@ -16,11 +16,11 @@ import com.sunya.cdm.api.computeSize
 class MaxChunker(val maxElems: Int, val wantSpace: IndexSpace, varshape : IntArray) : AbstractIterator<IndexSpace>() {
     val totalNelems = wantSpace.totalElements
     val rank = wantSpace.rank
-    val strider = IntArray(rank)
+    val strider = LongArray(rank)
     val odo = Odometer(wantSpace, varshape)
 
     init {
-        var accumStride = 1
+        var accumStride = 1L
         for (k in rank - 1 downTo 0) {
             strider[k] = accumStride
             accumStride *= wantSpace.shape[k]
@@ -50,7 +50,7 @@ class MaxChunker(val maxElems: Int, val wantSpace: IndexSpace, varshape : IntArr
         // always use the full length of the innermost dimension
         val chunkShape = IntArray(rank) { idx ->
             if (idx == rank - 1) shape[idx] else {
-                var size: Int = (maxElems / strider[idx])
+                var size: Int = (maxElems / strider[idx]).toInt()
                 size = if (size == 0) 1 else size
                 Math.min(size, shape[idx] - current[idx])
             }
